# fhir_server/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "APIs for KryonFire Server Version-1",
    "description": "This API Document contains the FHIR server api.",
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/resource/{resource_type}": {
      "post": {
        "tags": [
          "FHIR"
        ],
        "summary": "Create a resource",
        "parameters": [
          {
            "in": "path",
            "name": "resource_type",
            "type": "string",
            "enum": [
              "Patient",
              "Practitioner",
              "Organization",
              "Encounter",
              "ServiceRequest",
              "Appointment"
            ],
            "required": true,
            "description": "type of resource"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Resource object",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful response."
          }
        }
      },
      "get": {
        "tags": [
          "FHIR"
        ],
        "summary": "Returns a single resource",
        "description": "Returns a single resource",
        "operationId": "getSingleResource",
        "consumes": [
          "application/fhir+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "resource_type",
            "type": "string",
            "enum": [
              "Patient",
              "Practitioner",
              "Organization",
              "Encounter",
              "ServiceRequest",
              "Appointment"
            ],
            "required": true,
            "description": "type of resource"
          },
          {
            "name": "resource_id",
            "in": "query",
            "description": "ID of resource to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "put": {
        "tags": [
          "FHIR"
        ],
        "summary": "Update a resource or create new keys if it does not exist",
        "description": "Update a resource",
        "operationId": "updateResource",
        "parameters": [
          {
            "in": "path",
            "name": "resource_type",
            "type": "string",
            "enum": [
              "Patient",
              "Practitioner",
              "Organization",
              "Encounter",
              "ServiceRequest",
              "Appointment"
            ],
            "required": true,
            "description": "type of resource"
          },
          {
            "name": "resource_id",
            "in": "query",
            "description": "ID of resource to return",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Resource object",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    }
  }
}